refactor: split manager media gallery router into read and write modules
